requirements/requirements-base.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.17
Django==1.7.1
django-celery==3.1.16
django-cors-headers==1.0.0
django-extensions==1.4.9
django-haystack==2.4.1
django-redis==3.8.0
djangorestframework==3.0.1
djangorestframework-camel-case==0.2.0
jsonpatch==1.9
jsonpointer==1.10
kombu==3.0.33
psycopg2==2.5.4
pymarc==3.0.3
pysolr==3.2.0
python-dateutil==2.3
pytz==2015.7
redis==2.10.5
requests==2.9.1
six==1.10.0
ujson==1.33
Unipath==1.0
